Tighten att string requirements; disallow additionalProperties
