+++ test/test_data/validation/usj/bad_sid.json
@@ -12,7 +12,7 @@
       "type": "chapter",
       "marker": "c",
       "number": "1",
-      "sid": "GEN 1"
+      "sid": "GEN X"
     },
     {
       "type": "para",
