+++ test/test_data/validation/usj/extra_property.json
@@ -6,6 +6,7 @@
       "type": "book",
       "marker": "id",
       "code": "GEN",
+      "banana": "split",
       "content": []
     },
     {
